DBV-73 Allow duplicate chagnes
Liquibase fails to vaildate the change logs if they contain changes with
duplicate IDs, while this is generally correct, change logs may be
included several times, which will trigger this same failure, since the
changes are in fact the same change, not just sharing an ID, they should
be accepted. This change causes the validation to pass if the changes
which share an ID also share an MD5 sum.